requirements.txt
dagster==1.3.4
dagit==1.3.4
dagster-docker==0.19.4
dagster-gcp==0.19.4
dagster-postgres==0.19.4
pandas==2.0.1
google-cloud==0.34.0
requests==2.30.0
gcsfs==2023.5.0
pre-commit==3.3.1